Implement Field Registry API with Excel import/export support (#611)
* submit code for Filed Registry

submit code for Filed Registry

* Submit code for field Registry

Submit code for field Registry

* added get api for getfieldregistry.

* Excel import/export; remove Icon/ApiRoute

Add Excel import/export support and related helpers/controllers: introduce ExcelImportFormDto and ExcelImportHelper (ClosedXML), implement ExportPropertiesToExcelAsync and ImportPropertiesFromExcelAsync in CommonDetailsService with validation, identity resolution and transactional all-or-nothing updates, add file upload endpoint with file validation and rate limiting, and expose export endpoint. Update service interfaces to include import/export. Remove deprecated IconName and ApiRoute properties across DTOs, entity and DbContext, and adjust mappings/tests accordingly. Improve FieldRegistryService: add search filtering for schema/table queries, paging adjustments, and implement DeleteFieldRegistryAsync plus corresponding controller endpoint and test updates.

* copilot issues resolved

* Refactor FieldRegistry to support activate/deactivate

Add support for activating/deactivating field registries and refactor metadata lookup. IFieldRegistryService.DeleteFieldRegistryAsync was renamed to SetActiveStatusAsync and updated to accept an isActive flag; FieldRegistryService was rewritten to use EF Core model metadata (GetEntityTypes/GetProperties) instead of raw SQL to enumerate schemas, tables and columns, and now applies isActive to master and field configs. FieldRegistryController route changed from DELETE to PATCH and messages/logging updated accordingly. Also: add duplicate header detection in ExcelImportHelper, trim/validate wardNo/propertyNo in CommonDetailsService and include additional DTO fields; expand unit tests (import/export Excel and new FieldRegistryControllerTests) and small using/import adjustments.

* Add reassessment and retrospective tax APIs

Introduces new DTOs, service interfaces/implementations, and authorized controllers for Property Re-Assessment and Retrospective Tax screens, replacing legacy SQL/PIVOT behavior with EF Core LINQ workflows. Registers the new services in DI and wires retrospective tax data with required year/tax fields, foreign keys, and indexes in the EF model. Also fixes RV result type mismatches by mapping SQL float-backed fields as `double` across entities, tax-engine logic, and tests to prevent cast/runtime issues.

* Refactor BuildingPlanType key and same-as flow

Switch BuildingPlanType handling from PropertyId-based rows to a building-level key (WardId + PropertyNo + Type), including entity and EF mapping updates. DataEntrySameAs typewise upsert now inserts at most one building-level row after a building-level eligibility check, instead of inserting per property. The property fetch for same-as now joins PropertyDetails and returns aggregated CarpetAreaSqMeter/CarpetAreaSqFeet in the DTO. Also adds global.json to pin the .NET SDK version.

* Delete global.json

* copilot isses resolved

---------

+++ src/Application/NtisPlatform.Application/Services/TaxEngine/RateableValueCalculatorService.cs
@@ -149,25 +149,25 @@ namespace NtisPlatform.Application.Services.TaxEngine
                 PropertyId = detail.PropertyId,
                 PropertyDetailsId = detail.Id,
 
-                MonthlyRate = Math.Round(monthlyRate, 2),
-                YearlyRate = yearlyRate,
-                YearlyRent = yearlyRent,
+                MonthlyRate = (double)Math.Round(monthlyRate, 2),
+                YearlyRate = (double)yearlyRate,
+                YearlyRent = (double)yearlyRent,
 
                 Depreciation = depreciationAmount,
                 DepreciationPer = depreciationRate,
                 AppliedOn = appliedOn,
 
-                AnnualRentalValue = annualRentalValue,
+                AnnualRentalValue = (double)annualRentalValue,
                 Maintenance = maintenance,
                 RateableValue = rateableValue,
 
-                TotalAreaSqMtr = areaSqMtr,
+                TotalAreaSqMtr = (double)areaSqMtr,
 
                 RAreaSqMtr = string.Equals(typeOfUse.Type, "R", StringComparison.OrdinalIgnoreCase)
-                    ? areaSqMtr : 0m,
+                    ? (double)areaSqMtr : 0d,
 
                 CAreaSqlMtr = string.Equals(typeOfUse.Type, "C", StringComparison.OrdinalIgnoreCase)
-                    ? areaSqMtr : 0m
+                    ? (double)areaSqMtr : 0d
             };
         }
 
@@ -202,18 +202,18 @@ namespace NtisPlatform.Application.Services.TaxEngine
             {
                 PropertyId = detail.PropertyId,
                 PropertyDetailsId = detail.Id,
-                MonthlyRate = 0m,
-                YearlyRate = 0m,
-                YearlyRent = 0m,
+                MonthlyRate = 0d,
+                YearlyRate = 0d,
+                YearlyRent = 0d,
                 Depreciation = 0m,
                 DepreciationPer = 0m,
                 AppliedOn = appliedOn,
-                AnnualRentalValue = 0m,
+                AnnualRentalValue = 0d,
                 Maintenance = 0m,
                 RateableValue = 0m,
-                TotalAreaSqMtr = 0m,
-                RAreaSqMtr = 0m,
-                CAreaSqlMtr = 0m
+                TotalAreaSqMtr = 0d,
+                RAreaSqMtr = 0d,
+                CAreaSqlMtr = 0d
             };
         }
     }
